% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_parameters.R
\docType{data}
\name{default_parameters}
\alias{default_parameters}
\title{Default parameters for SEM.}
\format{A list with lots of entries:
\describe{
\item{price}{price, in US dollars}
\item{carat}{weight of the diamond, in carats}
...
}}
\usage{
default_parameters
}
\description{
Default parameters for SEM.
}
\keyword{datasets}
