% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_fxns.R
\name{format_inputs}
\alias{format_inputs}
\title{format_inputs}
\usage{
format_inputs(nc_path)
}
\arguments{
\item{nc_path}{The path to the Dietze and Matthes 2014 netcdf file or a file with the same structure.}
}
\value{
a data frame consisting of columns of the following:
\describe{
\item{time}{the day of the year}
\item{PAR}{incoming photosynthetically active radiation, umol/m2/s}
\item{temp}{air temperature, degrees C}
\item{VPD}{vapor pressure deficit, kPa}
\item{precip}{precipitation, mm}
}
}
\description{
Format the netcdf file form the Dietze and Matthes 2014 manuscript into a data frame that can bse used by \code{iterate.SEM}
}
