% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{iterate.SEM}
\alias{iterate.SEM}
\title{Solve SEM}
\usage{
iterate.SEM(
  pest,
  inputs,
  params = default_parameters,
  t.start = 7000,
  years = 1
)
}
\arguments{
\item{pest}{\link{phloem, xylem, leaf, root, stem} a vector of pest impacts, each entry in the vector
represents the percent change in thepholme, yxlem, leaf, root, and stem based on the type of disruption.}

\item{inputs}{the data.frame object returned by \code{format_inputs}}

\item{params}{a list of parameters, the default is set to the default parameters}

\item{t.start}{documenation needed}

\item{years}{documentation needed}

\item{X}{= \link{leaf,wood,root,storage,som,SoilWater,stem density}}
}
\description{
Solve SEM
}
