% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PestEcosysDemo.R
\name{SEM}
\alias{SEM}
\title{Simple Ecosystem Model}
\usage{
SEM(X, params, inputs, pest = c(0, 0, 0, 1, 0), timestep = 1800)
}
\arguments{
\item{X}{= \link{leaf,wood,root,storage,som,SoilWater,stem density}}

\item{params}{params}

\item{pest}{\link{phloem, xylem, leaf, root, stem}}

\item{timestep}{is in seconds, defaults to 30 min}

\item{inputs:}{PAR, temp, VPD}
}
\value{
X
}
\description{
Simple Ecosystem Model
}
\author{
Michael C, Dietze \href{mailto:dietze@bu.edu}{dietze@bu.edu}
}
