% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{SEM}
\alias{SEM}
\title{Simple Ecosystem Model}
\usage{
SEM(X, params, timestep, inputs, pest = c(0, 0, 0, 1, 0), p = 1800)
}
\arguments{
\item{X}{array \link{leaf,wood,root,storage,som,SoilWater,stem density}}

\item{params}{params}

\item{timestep}{is in seconds, defaults to 30 min}

\item{inputs}{PAR, temp, VPD}

\item{pest}{\link{phloem, xylem, leaf, root, stem} a vector of pest impacts, each entry in the vector
represents the percent change in thepholme, yxlem, leaf, root, and stem based on the type of disruption.}

\item{p}{p}
}
\value{
X
}
\description{
Simple Ecosystem Model
}
\author{
Michael C, Dietze \href{mailto:dietze@bu.edu}{dietze@bu.edu}
}
